// packages/sn-table/src/types.ts
export type Align = "left" | "center" | "right";
// ...
export interface Cell {
  qText?: string;
  qNum?: number | string;
  align: Align;
  qAttrExps?: EngineAPI.INxAttributeExpressionValues;
  qElemNumber: number;
  rowIdx: number;
  colIdx: number;
  pageRowIdx: number;
  pageColIdx: number;
  selectionColIdx: number;
  isSelectable: boolean;
  isLastRow: boolean;
  isLastColumn?: boolean;
  isNumeric?: boolean;
}

export interface Row {
  // for the row key, string is needed
  [key: string]: Cell | string;
}
// ...
export interface ExtendedSelectionAPI extends stardust.ObjectSelections {
  on(eventType: string, callback: () => void): void;
  removeListener(eventType: string, callback: () => void): void;
}
// ...
export interface AnnounceArgs {
  keys: Array<string | Array<string>>;
  shouldBeAtomic?: boolean;
  politeness?: "polite" | "assertive" | "off";
}

export type Announce = (arg0: AnnounceArgs) => void;